requirements.txt
streamlit
google-genai
python-dotenv
scikit-learn